feat: editing dtos

+++ src/main/java/br/com/luizalabs/verticallogistica/core/domain/dto/ProductsDto.java
@@ -2,5 +2,5 @@ package br.com.luizalabs.verticallogistica.core.domain.dto;
 
 import java.math.BigDecimal;
 
-public record ProductsDto(int product_id, BigDecimal value) {
+public record ProductsDto(long product_id, BigDecimal value) {
 }
