+++ src/main/java/br/com/luizalabs/verticallogistica/core/domain/dto/UserDto.java
@@ -2,5 +2,5 @@ package br.com.luizalabs.verticallogistica.core.domain.dto;
 
 import java.util.List;
 
-public record UserDto(int user_id, String name, List<OrderDto> orders) {
+public record UserDto(long user_id, String name, List<OrderDto> orders) {
 }
